Added file swap_strips.py
Swaps 2 strips between them. More specific, places the first strip in the channel and starting frame (frame_final_start) of the second strip, and places the second strip in the channel and starting frame (frame_final_end) of the first strip. If the biggest in duration strip doesn't fit in the space of the smallest strip, it does nothing.