Added operator to swap 2 strips and some utils (#185)
* Added file move_final_start.py

This file contains a module to move a strip at a specific frame, based on its frame_final_duration, without changing its duration.

* Added file swap_strips.py

Swaps 2 strips between them. More specific, places the first strip in the channel and starting frame (frame_final_start) of the second strip, and places the second strip in the channel and starting frame (frame_final_end) of the first strip. If the biggest in duration strip doesn't fit in the space of the smallest strip, it does nothing.

* Added file swap_strips_op.py

Swaps the 2 selected strips between them. More specific, places the first strip in the channel and starting frame (frame_final_start) of the second strip, and places the second strip in the channel and starting frame (frame_final_end) of the first strip. If the biggest in duration strip doesn't fit in the space of the smallest strip, it does nothing.

* Rename swap_strips_op.py to operators/swap_strips_op.py