Changed swap_strips.py
The previous approach of checking if a swap was possible, and the taking the correct action, was prove to bugs, because of the many special cases. Instead a new approach has been implemented. The swap takes places, without checking its possibility, and then we check if the strips are in a consistent state. If thats the case, then the swap was made successfully, otherwise, the swap isn't possible, so it gets cancelled, and the strips reset to their initial location without any changes.