Changed select_related_strips.py
The operator now selects all related strips to the selection. It also works with effect strips as the initial selection, with chains of effects on effects and effects with 2 inputs (i.e. cross).

How it works: Strips and its effects define a graph (we can consider it as an undirected one). Every strip is a node, while an edge is the connection of a strip with its effect (i.e. base_strip -> effect_strip, '->' is the edge, while base_strip and effect_strip the nodes). Using the bpy.ops.transform.seq_slide operator and with the input strips of the effects we can find all the neightbours of a strip in the graph, so we can apply a Depth First Search traversal to find all the related strips.

*Note for method find_neighbours: As stated in its documentation, it finds more neighbours than asked, but because extra code is needed to filter those, for brevity I permited that. The algorithm stays correct even with this behavior, the extra neighbours find_neighbours finds, are at most redundant. As of the performance in total, it is not clear if this behaviour of find_neighbours adds overhead or makes the algorithm faster, of course all that by a constant factor, the space and time complexities stay the same in both cases (not that it matters anyway).