Fixed a bug in snap_selection_to_cursor.py
Now it can more consistently snap a selection to cursor. Also, support for effect strips has been added, they can translate along with their strips.

Note: It is in the responsibility of the user that the snap is possible, i.e. there are no other blocking strips in the way. If that's not the case, then this operator will have an undefined behavior.