WIP Extract shortcut/docs from operator classes
A first idea of how to extract information from operators to generate
docs and shortcuts in the `json` format. Moved all info into a variable
called `doc` in the operators, the `description` entry in the `doc`
dictionary can for example be `__doc__` (the class documentation). This
is the simplest solution I found. I've done this only for `CrossfadeEdit`
as an example (doesn't have a shortcut cause it isn't registered with a
shortcut at the time of this writing).

The python script to generate the `json` file is very simple and it's in
`scripts/ShortcutsDocs`. It has to be ran through Blender (because of
`bpy` and other Blender modules), hence the
`scripts/ShortcutsDocs/shortcuts_docs.sh` script.

Extra shortcuts (that aren't linked to power sequencer operators) are
stored in `doc` inside of `operators/__init__.py`. So far works pretty
well, but I can't test the readme generation, cause I don't have hugo