Automate shortcuts and docs
All operator names and docs are stored in a `doc` class variable in the
operator. In order for new operators to work they have to have this `doc`
variable and it should be in the same format as current operators have:

```python
doc = {
    'name': doc_name(__qualname__),
    'demo': 'https://i.imgur.com/ZyEd0jD.gif',
    'description': doc_description(__doc__),
    'shortcuts': [
        ({'type': 'PLUS', 'value': 'PRESS', 'shift': True}, {}, 'Add Speed')
    ],
    'keymap': 'Sequencer'
}
bl_idname = doc_idname(doc['name'])
bl_label = doc['name']
bl_description = doc_brief(doc['description'])

```

Most of it is autogenerated using `__qualname__` (class name) & `__doc__`.
The `'shortcuts'` entry is an array (which can be empty) of tuples with 3 items:

1st item. is a dictionary that can be passed to `keymap_items.new()`
          function via `**` operator
2nd item. is a dictionary defining properties for the given shortcut
3rd item. is a description string of the functinality

closes #157