fade_add: try to preserve existing fade animation as best as possible
fade_animation_clear now clears animation in the fades' range instead of
deleting the entire fade animation and re-creating it from scratch.

This required calculating the fades' coordinates for each sequence.

I've used the operator's properties and mutate them for every sequence it loops
over to make it easy to split it into self-describing methods, but a creating
and passing a Fade object around might help with clarity and avoiding bugs