Update functions.py
It's not exactly a bug, but when trimming a a strip bigger than the one moved, all of the strips are deselected. What I do here is to reselect by saving the tuple, and the pass a for loop at the end.
I thought it would be better do it here, as it's here where it's being deselected, not in the channel_offset.py (I found out this issue while working in this file).