Refactor channel_offset.py
This solves a bug of wrongly trimming and deleting when doing it with 2 or more strips selected.

To trigger it, create 3 strips in 3 channels like:
3:         ██████████     ◄
2:              ██                     ◄
1: ██████████
Select the 2 from above and trim them down (ctrl + alt + down_arrow). The result is not the espected:
3: 
2:         ██████████
1: ███     ██

To solve the issue, I had to refactor the code and make the changes strip by strip, instead of block by block.
This allows to operate each strip individually and to set personalized range of trimming for each bit.
Since it now operates with changing ranges, I had to import all the operations from external files.
I know it's not "the blender way", but I was afraid to break compatibility.

I have also applied all changes from the pull requests done by me yesterday (crash when trimming down in channel 1,
upper limit, and disable the trimming when using alt + up/down_arrow).
Of course I also added the fixes for the function.py pull, since I had to import all functionalities.