Update channel_offset.py (#443)
* Update channel_offset.py

This fixes 2 issues:

The "trim_target_channel" change is because when using "ctrl+alt", and then using "alt" only, it doesn't disable the trimming operatioin.
The second (at line 88) solves a crash, where the strip is autodeleted when using "ctrl+alt+down_arrow" to a strip in channel 1. (More info in https://developer.blender.org/T82942, where I wrongly reported there).

* Update channel_offset.py

There was a 3rd critical issue here, where no upper limit was set, allowing for the strips to surpass the channel 32, and when hitting play, it crashed.