Update functions.py (#442)
* Update functions.py

It's not exactly a bug, but when trimming a a strip bigger than the one moved, all of the strips are deselected. What I do here is to reselect by saving the tuple, and the pass a for loop at the end.
I thought it would be better do it here, as it's here where it's being deselected, not in the channel_offset.py (I found out this issue while working in this file).

* Update functions.py

The strings product of the trimming from an upper string were not selected if the trimmed strip was. This adds to the list to reselect the remaining strips of the trimmed strip if this trimmed one was selected when doing the operation.

* Update functions.py

This fixes an odd behaviour when pushing the strips up in the 32 channel, where the order the context.sequences has of the strips are from 1 to 32 and then chronologically. That meant the reselection was incorrect, as the context.sequences[0] wasn't the last strip created, but one in a lower channel