Refactor channel_offset.py (#444)
* Refactor channel_offset.py

This solves a bug of wrongly trimming and deleting when doing it with 2 or more strips selected.

To trigger it, create 3 strips in 3 channels like:
3:         ██████████     ◄
2:              ██                     ◄
1: ██████████
Select the 2 from above and trim them down (ctrl + alt + down_arrow). The result is not the espected:
3: 
2:         ██████████
1: ███     ██

To solve the issue, I had to refactor the code and make the changes strip by strip, instead of block by block.
This allows to operate each strip individually and to set personalized range of trimming for each bit.
Since it now operates with changing ranges, I had to import all the operations from external files.
I know it's not "the blender way", but I was afraid to break compatibility.

I have also applied all changes from the pull requests done by me yesterday (crash when trimming down in channel 1,
upper limit, and disable the trimming when using alt + up/down_arrow).
Of course I also added the fixes for the function.py pull, since I had to import all functionalities.

* Update operators/channel_offset.py

Code optimization

Co-authored-by: Nathan Lovato <nathan@gdquest.com>

* Update channel_offset.py

I removed some redundant code in comparisons, added two functions and fixed the descriptions according to the changes.

* Update channel_offset.py

Some more optimizations

* Update functions.py

- Made consistent comments.
- Fixed the function find_strips_in_range, where in edge case strips where wrongly assigned to the strips_overlapping_range list
- Improved functionality of trim_strips, where keeps the selection of the current strips, and adds if new strips are created from one selected.
- Rewrite of the slice_selection, which was innecessary confusing and ineficient.
- Added move_selection, which cleans code from ripple_move and apply_time_offset. Also used for channel_offset.
- Added zoom, which can add depth to the slice_selection operation by giving current frame/pixel relation.

* Rework for channel_offset.py

- Rolled back to the blocks model.
- Added meaning for the blocks by adding an intelligent slicing method with zoom.
- Added the option to keep all the selection's offset between each other while moving.
- Optimized moving process by calling move_selected in ..\utils\function.py
The zoom function, if not wanted, can be easily deleted from the code.
The keep_selection_offset is set here to be True by default. Feel free to change it if.

* Update functions.py

I forgot to include a major change to slice_selection. I refactored all the code in there as it was a pain to understand and I figured out a simpler way to get the same result. I also added a 3rd parameter essential to make the blocks.

* Bug fix

* Rename of zoom() to pixel_frame_ratio()

* Rename of zoom() to pixel_frame_ratio()

* Little bug fix

* Rename pixel_frame_ratio

* Rename of pixel_frame_ratio

* Rename of calulate_pixel_frame_ratio

I thought it'd have more sense for the name to have the correct order of the relation

* Roll back frame_pixel_ratio (deletion)

* Roll back frame_pixel_ratio & optimizations

* Tweak code, fix typo, rename variables

* Fix error introduced in Snap Selection

Co-authored-by: Nathan Lovato <nathan@gdquest.com>